Add a delete confirmation dialog

+++ src/app/editor-wrapper/editor-wrapper.component.ts
@@ -3,14 +3,16 @@ import { NgxIndexedDBService } from 'ngx-indexed-db';
 import { Editor } from '../editor';
 import { MatSelectionList, MatSelectionListChange } from '@angular/material/list';
 import { IOutputData, SplitComponent } from 'angular-split';
+import { MatDialog } from '@angular/material/dialog';
+import { DeleteDialogComponent } from '../delete-dialog/delete-dialog.component';
 
-interface Item{
+interface Item {
     name: string;
     id: number;
     data: any;
 }
 
-interface ListItemData extends Item{
+interface ListItemData extends Item {
     index: number;
 }
 
@@ -32,10 +34,10 @@ export class EditorWrapperComponent implements AfterViewInit {
 
     @Input()
     public editor!: Editor;
-    
+
     public items: ListItemData[] = [];
 
-    @ViewChild("itemsList", {static: true})
+    @ViewChild("itemsList", { static: true })
     private itemsList!: MatSelectionList;
 
     @ViewChild("split", { static: true })
@@ -43,19 +45,19 @@ export class EditorWrapperComponent implements AfterViewInit {
 
     private itemsIndex = 0;
 
-    constructor(private dbService: NgxIndexedDBService, private changeDetector: ChangeDetectorRef){
+    constructor(private dbService: NgxIndexedDBService, private changeDetector: ChangeDetectorRef, public dialog: MatDialog) {
     }
 
-    ngAfterViewInit(): void{
-        if(this.storeName.length !== 0){
+    ngAfterViewInit(): void {
+        if (this.storeName.length !== 0) {
             this.dbService.getAll<Item>(this.storeName)
-            .subscribe((entries: Item[]) => {
-                this.items = entries.map((item, index) => ({...item, index: index}));
-            });
+                .subscribe((entries: Item[]) => {
+                    this.items = entries.map((item, index) => ({ ...item, index: index }));
+                });
         }
 
         const toolsSplitJsonData = localStorage.getItem("toolsSplitData");
-        if(toolsSplitJsonData){
+        if (toolsSplitJsonData) {
             this.toolsSplitElement.setVisibleAreaSizes(JSON.parse(toolsSplitJsonData));
             this.changeDetector.detectChanges();
         }
@@ -63,9 +65,9 @@ export class EditorWrapperComponent implements AfterViewInit {
         this.editor.onChange.subscribe((data) => this.onDataChange(data));
     }
 
-    
 
-    public add(){
+
+    public add() {
         this.itemsList.selectedOptions.selected[0]
         const item = {
             name: "New " + this.name + (this.itemsIndex++).toString(),
@@ -73,52 +75,62 @@ export class EditorWrapperComponent implements AfterViewInit {
         };
         this.dbService.add(this.storeName, item)
             .subscribe((record) => {
-                this.items.unshift({...record, index: 0});
+                this.items.unshift({ ...record, index: 0 });
                 this.updateIndices();
             });
     }
 
-    public deleteItem(index: number){
+    public deleteItem(index: number) {
+
         const item = this.items[index];
-        this.dbService.deleteByKey(this.storeName, item.id)
-            .subscribe((status) => {
-                this.items.splice(index, 1);
-                this.updateIndices();
-            });
+        const dialogRef = this.dialog.open(DeleteDialogComponent, {
+            data: { name: item.name },
+        });
+
+        dialogRef.afterClosed().subscribe(shouldDelete => {
+            if (shouldDelete){
+                this.dbService.deleteByKey(this.storeName, item.id)
+                    .subscribe((status) => {
+                        this.items.splice(index, 1);
+                        this.updateIndices();
+                    });
+            }
+        });
+        
     }
 
-    public rename(index: number, newName: string){
+    public rename(index: number, newName: string) {
         this.items[index].name = newName;
         this.updateCurrent(this.items[index]);
     }
 
-    protected onSelectionChange(event: MatSelectionListChange){
+    protected onSelectionChange(event: MatSelectionListChange) {
         this.editor.setData(event.options[0].value.data);
     }
 
-    protected onToolsSplitterDragEnd(data: IOutputData){
+    protected onToolsSplitterDragEnd(data: IOutputData) {
         localStorage.setItem("toolsSplitData", JSON.stringify(data.sizes));
     }
 
-    private onDataChange(data: any){
+    private onDataChange(data: any) {
         console.log("data change");
         const itemData = this.itemsList.selectedOptions.selected[0].value;
         this.items[itemData.index].data = data;
         this.updateCurrent(itemData);
     }
 
-    private updateCurrent(item: Item){
+    private updateCurrent(item: Item) {
         this.dbService.update(this.storeName, {
             name: item.name,
             id: item.id,
             data: item.data
         })
             .subscribe((value) => {
-                
+
             });
     }
 
-    private updateIndices(){
+    private updateIndices() {
         for (let i = 0; i < this.items.length; i++) {
             const item = this.items[i];
             item.index = i;
